Изменен тип данных в методе Factorial с int на long для расчета факториала больших чисел

+++ Seminar9/Program.cs
@@ -3,7 +3,7 @@
 Console.WriteLine("Введите число, факториал которого необходимо вычислить:");
 int value = Convert.ToInt32(Console.ReadLine());
 Console.WriteLine();
-int Factorial(int n)
+long Factorial(long n)
 {
 if (n == 0 || n == 1)
 return 1;
